starts filling in binance funcs

+++ exchanges/binance/cfutures_types.go
@@ -441,7 +441,7 @@ type FuturesPositionInformation struct {
 	MarkPrice        float64 `json:"markPrice,string"`
 	UnrealizedProfit float64 `json:"unRealizedProfit,string"`
 	LiquidationPrice float64 `json:"liquidation,string"`
-	Leverage         int64   `json:"leverage"`
+	Leverage         float64 `json:"leverage"`
 	MaxQty           float64 `json:"maxQty"`
 	MarginType       string  `json:"marginType"`
 	IsolatedMargin   float64 `json:"isolatedMargin,string"`
